requirements.txt
fastapi
pystray
customtkinter
packaging
python-multipart
aiofiles
requests
qrcode
zeroconf
uvicorn[standard]
pillow
mss
pyautogui
pynput
evdev; platform_system == "Linux"
psutil